requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
astroid==2.15.8
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cryptography==41.0.3
cssutils==2.7.1
cycler==0.11.0
dataframe-image==0.2.1
defusedxml==0.7.1
dill==0.3.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastjsonschema==2.18.0
fonttools==4.42.1
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
html2image==2.0.4.3
htmlwebshot==0.1.2
httplib2==0.22.0
hypothesis==6.87.1
idna==3.4
isort==5.12.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
mistune==3.0.1
multidict==6.0.4
mypy==1.5.1
mypy-extensions==1.0.0
mysqlclient==2.2.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
numpy==1.25.2
oauthlib==3.2.2
openapi-client @ file:///home/kevin/Coding/Learning/Projects/nutrition_bot/spoonacular_api
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
Pillow==10.0.0
platformdirs==3.10.0
protobuf==4.24.3
py-find-injection==0.1.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pylint==2.17.7
pyparsing==3.0.9
pyTelegramBotAPI==4.13.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.2
rsa==4.9
seaborn==0.12.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
telebot==0.0.5
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
webencodings==0.5.1
websocket-client==1.6.2
wrapt==1.15.0
yarl==1.9.2
